requirements.txt
asgiref==3.8.1
Django
django-cors-headers==4.4.0
django-gunicorn==0.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
uritemplate==4.1.1
